+++ dbldatagen/text_generator_plugins.py
@@ -63,6 +63,7 @@ class PyfuncText(TextGenerator):  # lgtm [py/missing-equals]
             of the `initFn` calls
 
             :param txtGen: - reference to outer PyfnText object
+
         """
 
         def __init__(self, txtGen):
@@ -185,7 +186,8 @@ class PyfuncTextFactory:
     def __init__(self, name=None):
         """
 
-        :param name:
+        :param name: name of generated object (when converted to string via ``str``)
+
         """
         self._initFn = None
         self._rootProperty = None
@@ -373,6 +375,7 @@ def fakerText(mname, *args, _lib=None, _rootClass=None, **kwargs):
        :param args: positional args to be passed to underlying Faker instance
        :param _lib: internal only param - library to load
        :param _rootClass: internal only param - root class to create
+       
        :returns : instance of PyfuncText for use with Faker
 
        ``fakerText("sentence")`` is same as ``FakerTextFactory()("sentence")``